Generate comparison data moved to relatedSlice, cards now create new modal
